requirements.txt
streamlit
mysql-connector-python
python-dotenv
toml
